reorg and add new helper methods
